陽性者数（区市町村別）: Rename components/ConfirmedCasesByMunicipalitiesTable.vue to components/index/CardsReference/ConfirmedCasesByMunicipalities/Table.vue
